requirements-dev.txt
-r requirements.txt
black==20.8b1
flake8-bugbear==20.1.4
flake8==3.8.4
freezegun==1.0.0
python-semantic-release==7.3.0
syrupy==1.0.0
SQLAlchemy-Utils==0.36.8
pytest-asyncio==0.14.0
asynctest==0.13.0
-e .[dev]
